Remove encoding, combine it with VTable struct
