+++ vortex-python/src/arrays/builtins/mod.rs
@@ -14,7 +14,8 @@ pub(crate) use primitive::*;
 use pyo3::prelude::*;
 pub(crate) use struct_::*;
 use vortex::arrays::{
-    BoolVTable, ExtensionVTable, FixedSizeListVTable, NullVTable, VarBinVTable, VarBinViewVTable,
+    BoolVTable, ExtensionVTable, FixedSizeListVTable, ListVTable, NullVTable, VarBinVTable,
+    VarBinViewVTable,
 };
 use vortex::encodings::bytebool::ByteBoolVTable;
