+++ vortex-python/src/arrays/py/mod.rs
@@ -2,10 +2,9 @@
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
 mod array;
-mod encoding;
 mod python;
 mod vtable;
 
 pub(crate) use array::*;
-pub(crate) use encoding::*;
 pub(crate) use python::*;
+pub(crate) use vtable::*;
